updated vocab function
