+++ wookie/__init__.py
@@ -1,3 +1,4 @@
 from wookie.connectors import showpairs, separatesides, createsbs
-from wookie.lrcomparators import LrDuplicateFinder, LrPruningConnector, LrTokenComparator, LrExactComparator
+from wookie.lrcomparators import LrDuplicateFinder, LrPruningConnector, LrExactComparator
 from wookie.sbscomparators import PipeSbsComparator, FuzzyWuzzySbsComparator, DataPasser
+from wookie.tokenizers import LrTokenComparator
